+++ src/components/Form/CardPasswordForm.tsx
@@ -1,7 +1,8 @@
 import FormElement from "../common/FormElement";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import Input from "./Input";
 import { CardFormProps } from "./CardNumberForm";
+import useCardPasswordForm from "../../hooks/useCardPasswordForm";
 
 interface CardPasswordFormProps extends CardFormProps {
   cardPassword: string[];
@@ -18,35 +19,12 @@ const CardPasswordForm = ({
   onValidation,
   onFocus,
 }: CardPasswordFormProps) => {
-  const [errorMessage, setErrorMessage] = useState("");
-  const [inputValidity, setInputValidity] = useState({
-    0: false,
-  });
   const [isFocused, setIsFocused] = useState(false);
-
-  const handleFocus = (type: string) => {
-    onFocus(type);
-    setIsFocused(true);
-  };
-
-  // NOTE: 각 입력 필드의 유효성 검사 결과를 업데이트
-  const updateInputValidity = (index: number, isValid: boolean) => {
-    setInputValidity((prevValidities) => ({
-      ...prevValidities,
-      [index]: isValid,
-    }));
-  };
-
-  useEffect(() => {
-    if (onValidation && inputValidity[0]) {
-      onValidation(true);
-      setErrorMessage("");
-    }
-    if (onValidation && !inputValidity[0]) {
-      onValidation(false);
-      setErrorMessage("2자리 수를 입력하세요.");
-    }
-  }, [inputValidity]);
+  const { errorMessage, setErrorMessage, updateInputValidity, handleFocus } = useCardPasswordForm({
+    onFocus,
+    setIsFocused,
+    onValidation,
+  });
 
   const inputs = Array.from({ length: inputCount }, (_, index) => (
     <Input
